requirements.txt
fastapi
uvicorn
tflite-runtime
python-multipart
pillow
numpy<2.0.0
groq
httpx
joblib
pandas
scikit-learn
pydantic
python-dotenv
